+++ reference_data/models.py
@@ -161,6 +161,16 @@ class GeneCopyNumberSensitivity(models.Model):
         json_fields = ['pHI', 'pTS']
 
 
+class GeneShet(models.Model):
+    gene = models.ForeignKey(GeneInfo, on_delete=models.CASCADE)
+
+    shet = models.FloatField()
+    shet_constrained = models.BooleanField()
+
+    class Meta:
+        json_fields = ['shet', 'shet_constrained']
+
+
 class Omim(models.Model):
     MAP_METHOD_CHOICES = (
         ('1', 'the disorder is placed on the map based on its association with a gene, but the underlying defect is not known.'),